requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

boto3==1.42.84
fastapi==0.135.3
httpx==0.28.1
kaggle==2.0.0
mangum==0.21.0
matplotlib==3.10.8
numpy==2.4.4
pandas==3.0.2
pydantic==2.12.5
pytest==9.0.2
PyYAML==6.0.3
requests==2.33.1
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
statsmodels==0.14.6
uvicorn==0.41.0
xgboost==3.2.0
